Text can now be shared with Fulguris
Provided text is just put in the address bar.
Thus giving the user the opportunity to modify it before using it.
We also had to simplify tabs setup on app launch from intent by using
the same code as when a new intent is received while the app is already running.
Close Slion/Fulguris#279.