Replace deprecated ButterKnife with view binding
Is this really necessary?
View Binding has been supported for a long time and is actively maintained.

ButterKnife is outdated and obsolete.

Existing versions of ButterKnife will of course continue to work, but only critical bug fixes for integration with AGP will be considered.
Development of features and general bug fixes has been discontinued.
From there I see the point of removing this library.