Exporting the History fully works!
Exporting the history to JSON is fully working, first try! :)
I mostly copied what exporting Bookmarks does, which meant I had to add a sealed History class with a Entry data class to WebPage, which caused problems in SuggestionsAdapter's "when (webPage)", which I fixed by adding it also as an (unused) "is History" case, which shouldn't make any problems down the road hopefully.

Perhaps it could be cleaned up a little bit, with better names for example.

Next step would be to allow importing the history from a backup, a private addHistoryEntry function already exists which I could reuse for this. What has to be decided first is do I just replace the history (after warning the user) or do I offer a choice between replacing it or appending the history to the already existing one (this could be done by exporting both, combining the JSONs and importing that so it's not that important).