Options menu
New options menu displayed in bottom sheet.
It currently features domain and scrollbars settings.
It's designed to host settings that can have an immediate effects on the browser activity.
For instance as you adjust the scrollbars size you can see it being applied in the WebView underneath.
Settings that apply to the current tab such as domain settings are also good options candidates.
Related to #497 and #159.